+++ src/components/CreateQuizButton/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button } from '@mantine/core';
+import { Affix, Button } from '@mantine/core';
 import Icon from '../../components/Icon';
 import { useNavigate } from 'react-router';
 
@@ -11,27 +11,23 @@ type Props = {
 function CreateQuizButton({ userName, isFloating = false }: Props): JSX.Element {
   const navigate = useNavigate();
 
-  return (
+  const Btn = (
     <Button
       size={isFloating ? 'xl' : 'lg'}
       onClick={() => navigate(`/configure-quiz/${userName}`)}
       variant="filled"
-      sx={(theme) => ({
-        boxShadow: theme.shadows.sm,
-        ...(isFloating
-          ? {
-              position: 'fixed',
-              bottom: '30px',
-              right: '30px',
-              boxShadow: '2px 2px 12px -2px var(--shadow-color-dark)',
-            }
-          : {}),
-      })}
+      sx={(theme) => ({ boxShadow: isFloating ? theme.shadows.lg : theme.shadows.sm })}
       radius="xl"
       leftIcon={<Icon name="plus" color="#ffffff" width={20} height={20} />}>
       Create Quiz
     </Button>
   );
+
+  if (isFloating) {
+    return <Affix position={{ bottom: 30, right: 30 }}>{Btn}</Affix>;
+  }
+
+  return Btn;
 }
 
 export default CreateQuizButton;
